// frontend/src/api/generated/models/bookmark-property-value-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * chainlink-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookmarkPropertyValueJson
 */
export interface BookmarkPropertyValueJson {
    /**
     * 
     * @type {string}
     * @memberof BookmarkPropertyValueJson
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkPropertyValueJson
     */
    definitionId: string;
    /**
     * 
     * @type {string}
     * @memberof BookmarkPropertyValueJson
     */
    valueText?: string;
    /**
     * 
     * @type {number}
     * @memberof BookmarkPropertyValueJson
     */
    valueNumber?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BookmarkPropertyValueJson
     */
    valueBoolean?: boolean;
}

/**
 * Check if a given object implements the BookmarkPropertyValueJson interface.
 */
export function instanceOfBookmarkPropertyValueJson(value: object): value is BookmarkPropertyValueJson {
    if (!('definitionId' in value) || value['definitionId'] === undefined) return false;
    return true;
}

export function BookmarkPropertyValueJsonFromJSON(json: any): BookmarkPropertyValueJson {
    return BookmarkPropertyValueJsonFromJSONTyped(json, false);
}

export function BookmarkPropertyValueJsonFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookmarkPropertyValueJson {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'definitionId': json['definitionId'],
        'valueText': json['valueText'] == null ? undefined : json['valueText'],
        'valueNumber': json['valueNumber'] == null ? undefined : json['valueNumber'],
        'valueBoolean': json['valueBoolean'] == null ? undefined : json['valueBoolean'],
    };
}

export function BookmarkPropertyValueJsonToJSON(json: any): BookmarkPropertyValueJson {
    return BookmarkPropertyValueJsonToJSONTyped(json, false);
}

export function BookmarkPropertyValueJsonToJSONTyped(value?: BookmarkPropertyValueJson | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'definitionId': value['definitionId'],
        'valueText': value['valueText'],
        'valueNumber': value['valueNumber'],
        'valueBoolean': value['valueBoolean'],
    };
}